PGN support and minor improvements
The board now tracks all moves made since it was created. It can also convert this list of moves into PGN notation, which can be moved into other chess games or simply used to post games. There is lots more that can be done here. This will be the format we use to save games.