Reworks to move validation
The knight move is so simple that it's method has been removed. I added a method for diagonal move checks to shorten the code for bishops and queens, and make it more readable. If we like this change I will do the same for horizontal moves in the rook and queen.